refactor() Reviewing conversion to avoid out of memory error
