+++ src/core/services/loading/loading_old/augmented_strategies.py
